fix(i18n): move formatError into i18n-error.ts so its test avoids trpc.ts's DB import
trpc.ts imports @superset/db/client at module scope, which calls neon()
at import time; CI's test job has no DATABASE_URL, so importing
formatError from trpc.ts crashed the suite there (masked locally by
.env). The formatter belongs with the contract module anyway; trpc.ts
re-exports it.
